moved public functions out from `Workspace` and into `PC`
